backend/database/supabase_client.py
"""
Supabase Database Client
"""
from supabase import create_client, Client
import os
from dotenv import load_dotenv

load_dotenv()

# Get environment variables
SUPABASE_URL = os.getenv('SUPABASE_URL')
SUPABASE_KEY = os.getenv('SUPABASE_KEY')

if not SUPABASE_URL or not SUPABASE_KEY:
    raise ValueError("SUPABASE_URL and SUPABASE_KEY must be set in environment variables")

# Create client directly (simpler approach for newer supabase versions)
supabase: Client = create_client(SUPABASE_URL, SUPABASE_KEY)

